requirements.txt
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
coreapi==2.3.3
coreschema==0.0.4
cymem==2.0.8
dataclasses-json==0.6.6
datasets==2.19.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
filelock==3.14.0
FlagEmbedding @ git+https://github.com/FlagOpen/FlagEmbedding.git@ec3e54e4c0c67e9864057307996abaf867555107
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
intel-openmp==2021.4.0
itypes==1.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpath-ng==1.6.1
langcodes==3.4.0
language_data==1.2.0
llama-index==0.10.37
llama-index-agent-openai==0.2.5
llama-index-cli==0.1.12
llama-index-core==0.10.37
llama-index-embeddings-openai==0.1.9
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-groq==0.1.4
llama-index-llms-openai==0.1.19
llama-index-llms-openai-like==0.1.3
llama-index-multi-modal-llms-openai==0.1.6
llama-index-postprocessor-flag-embedding-reranker==0.1.3
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.22
llama-index-readers-llama-parse==0.1.4
llama-index-readers-remote==0.1.4
llama-index-readers-remote-depth==0.1.4
llama-index-readers-youtube-transcript==0.1.4
llama-index-vector-stores-pinecone==0.1.7
llama-parse==0.4.3
llamaindex-py-client==0.1.19
marisa-trie==1.1.1
MarkupSafe==2.1.5
marshmallow==3.21.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.30.1
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pinecone-client==3.2.2
ply==3.11
preshed==3.0.9
psutil==5.9.8
psycopg2==2.9.9
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
srsly==2.4.8
striprtf==0.0.26
sympy==1.12
tbb==2021.12.0
tenacity==8.3.0
thinc==8.2.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.40.2
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
whitenoise==6.6.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
youtube-transcript-api==0.6.2
